**Test appication UI and fix and update any errors and changes**
The goal of this commit is to test the UI and fix errors that have occurred

**Batch count not updating after 2FA login**

When a user authenticates with 2FA, the *used count* in the batch card does not increment as expected.

The problem lies in the cache. The paginator object that displays the batch card history which includes details such as number issued, expiry date, etc is also cached. As a result, it doesn't update until the cache expires. When the cache expires, the application pulls fresh data from the database, updates the paginator, and re-caches the paginator object for use. For the login page, this means that after a user logs in, the batch card history **will not** reflect the latest data unless the cache has already expired.

To resolve this, the application needs to be explicitly notified that an update has occurred. Instead of serving data from the cache, it should fetch the new information from the database pass it to the paginator and then re-cache the data including the paginator.

**Code change:**
```python
request.session["force_update"] = True
````

This ensures that whenever a user logs in with 2FA, the system bypasses the cache and pulls the latest batch card histories directly from the database, allowing the *used count* and other details to update immediately and then caches that data for future use.

**Frontend card history not reflecting invalidation**

When a code is entered in the *invalidate and delete* form, it is correctly invalidated in the backend. However, the batch card history on the frontend does not update in real-time without a page refresh, so users continue to see outdated information unless the page is refreshed. Since this a SPA (Single page Application) we don't want the user to always refresh to see the correct information

The issue was in the JavaScript function `updateCurrentRecoveryCodeBatchCard`.

Originally, the function took **two parameters**: the card element and the field to update (e.g., number of codes issued). Later, a new function `getCurrentCard()` was added to always select the current card internally without the need to pass it the current card, so the function only needed the **field name**.

However, the old first parameter (`recoveryBatchSectionElement`) was not removed from the function definition. Other parts of the code were calling the function with just the field name. **Since** JavaScript does **not** throw an error when a function is called with fewer arguments than expected, which is one of the most annoying aspect of JS, the function ran but did not update the card correctly.

**Original function:**
```javascript
updateCurrentRecoveryCodeBatchCard(recoveryBatchSectionElement, fieldName) {
  // ...update card code here
}
````

**Changed function to**
```
javascript
updateCurrentRecoveryCodeBatchCard(recoveryBatchSectionElement, fieldName) { <- forgot to remove the `recoveryBatchSectionElement`
  const currentCard = getCurrentCard()
  // ...update card code here
}

```

**Incorrect call:**

```javascript
updateCurrentRecoveryCodeBatchCard(fieldName) // only one argument which is getting passed as a `recoveryBatchSectionElement` parameter
```

**Fixed function:**

```javascript
updateCurrentRecoveryCodeBatchCard(fieldName) {
  const card = getCurrentCard();
  // ...update card code here
}
```

**Email/Download buttons not resetting after batch regeneration**

When a user clicks the **Email** or **Download** buttons, they are changed to *Emailed* and *Downloaded* as expected. However, after regenerating a batch, the buttons remain in that state and do not reset to allow new actions. Even refreshing the page does not fix it.

The issue is similar to Problem 1 and is caused by caching. When a user clicks the **Email** or **Download** button, the action is recorded in both the database and the cache. The cache is then used to mark the buttons as *Emailed* or *Downloaded*. On page refresh, the buttons’ states are pulled from the cache and not from the **database**, which in this case had not been updated after regenerating a batch.

To fix this, the session values that track the buttons’ states are reset, and the cache is instructed to update from the database.

**Code change:**
```python
request.session["is_downloaded"] = False
request.session["is_emailed"]    = False
request.session["force_update"]  = True
````

**Missing batch IDs in scheduler email reports**

When schedulers run, the batches listed in the email report do not include their IDs. Without IDs, it can appear that the same batch is repeated, even if multiple different batches were processed. This makes it difficult for the admin to distinguish which batches were actually included in the report.

1. The scheduler calls a task that in turn calls the function `purge_all_expired_batches_codes`.
2. This function retrieves all expired batches and loops through each one. Each batch instance calls its method `purge_expired_codes` to delete its own expired codes. The method returns the number of codes deleted and whether the batch is now empty.
3. The batch instance is then passed to `_generate_purged_batch_code_json_report`, during the loop which generates the report for the admin and then adds it to the main dictionary.
4. Since the batch is deleted, some data is still accessible (e.g., `batch.number_issued`) through the instance since it basically a variable (a snapshot of a given moment), but the **batch ID** becomes `None` even through the variable.

The method `purge_all_expired_batches_codes` was updated to also return the `batch id` **before** deletion as well. This ID is then passed to `_generate_purged_batch_code_json_report`, during the loop, ensuring that each batch in the report includes its ID and is uniquely identifiable.

Scheduler email reports now display batch IDs alongside their content, making it clear which batches were processed and preventing confusion about duplicate entries.
`

 ## Additional updates

- Added docstrings to admin models in the **admin.py**
- Added a message in the login page to inform the user that they can use 2FA Recovery Codes as backup

**utils.js**

**Modification:**
The function `checkIfCorrectHTMLElementHelper` was updated to include a **silent** parameter.
- `silent: true` → suppresses errors
- `silent: false` (default) → shows errors as usual

**Rationale:**
The application uses this helper to check whether elements are valid, either for a one-time check or dynamically when elements are used.

Dynamic elements like spinners do not appear in the DOM until certain conditions are met. Because the helper always checked them, it would log errors to the console unnecessarily.

**Solution:**
When calling `checkIfCorrectHTMLElementHelper` for dynamic elements (e.g., via `toggleSpinner`), we now pass `silent: true`. This suppresses errors while still allowing warnings, ensuring cleaner console output without affecting functionality.