**Fix:** Correct type enforcement error in scheduler
* The aim of this commit is to fix an error that was being thrown when the scheduler ran.

**Error details**
The error was raised by the `@enforce_types()` decorator:

```
TypeError: Argument `batch_id` expected type UUID got str
```

**What is `enforce_types`?**

`enforce_types` is a decorator that checks the types of parameters at runtime and raises an error if a type does not match the expected one.

**Problem**

The `enforce_types` decorator was raising an error because the method defined  parameter (`batch_id: str `) for the type hinting  but the batch_id is actually a UUID
not a string.

**Fix**

The fix was to import the uuid as a type and use as type hinting

**Previous**

```python
@enforce_types()
process_batch(self,
                      batch: RecoveryCodesBatch,
                      purged_count: int,
                      is_empty: bool,
                      batch_id: str,
                      use_with_logger: bool = True)

```

**Updated**

```python

from uuid import UUID

@enforce_types()
process_batch(self,
                      batch: RecoveryCodesBatch,
                      purged_count: int,
                      is_empty: bool,
                      batch_id: UUID, # <------------------- uuid type just like "int", etc
                      use_with_logger: bool = True)

```