**Feat:** Add conditional logic to decide when to hide context and headers
The goal of this commit is to introduce an `if` statement that determines when the **context** and **headers** should be logged.

Not all context information is sensitive, so the function calling the internal private method `_if_set_to_true_use_logger` can now decide, via the `log_context_and_header` flag, whether to include these fields in the logs. This gives calling function the discretion to log safely when appropriate.

---

When logging is enabled but the `context`, `headers`, and `log_context_and_header` flags are set to `False`, these fields are automatically excluded from the output.

Note how the log files now show the context and headers as *skipped* because they were never included in the log output.

Example:

```text
[INFO    ] email_sender: The following fields were logged: 'subject, html_template, to_email, text_template, from_email', and '2' were skipped. | category=SUMMARY | action=FIELDS_LOGGED_AND_SKIPPED
[INFO    ] email_sender: Number of fields logged: 5. | category=SUMMARY | action=FIELDS_LOGGED
[INFO    ] email_sender: Number of fields skipped: 2, fields skipped: 'context, headers'. | category=SUMMARY | action=FIELDS_SKIPPED
[INFO    ] email_sender: Optional fields are 'Context' and 'Headers'. Headers set: False, Context set: True. | category=SUMMARY | action=FIELDS_FULL_SUMMARY
[INFO    ] email_sender: End of summary log. | category=SUMMARY | action=END
```

---

This change adds conditional logging to `EmailSenderLogger`, allows sensitive information in the context or headers
 to be excluded based on the `log_context_and_header` flag and gives improved flexibility and security.