**Fix:** Verification test form not displaying correctly
**Goal**

The goal of this commit is to fix the issue where the test verification form does not appear after the initial recovery code generation, but only after a page refresh. The initial recovery code generation meaning when the user first time setup.

**How it works and expected result**

When a user generates their 2FA recovery code for the first time, a test verification form is displayed without a page refresh so that they can confirm their backend and frontend are correctly set up.

This form appears **only during the first setup**, meaning when the user creates their recovery code for the first time. It will remain visible until the user successfully verifies their code, after which it will no longer be shown, even if more codes are generated later.

**Problem**

Currently, when the user generates their recovery code for the first time, the verification form is not displayed automatically. It only appears after a manual page refresh, which is not the intended behaviour.

**Fix**

1. The issue was caused by an error triggered when the card was being created. The app attempted to access the length of an element that was `null`, which threw an error and prevented the rest of the script from running.

   A defensive `if` statement was added to check whether the `fieldElements` variable is empty. If it is, the function exits early. Otherwise, it proceeds to check the length of the elements.

   ```js
   // updatedBatchHistorySection.js

   export function markCardFieldHelper(cardElement, className, value, classColourSelector = "text-green") {
       const fieldElements = getCardFieldElements(cardElement);

       if (!fieldElements) return; // Defensive check added

       if (!fieldElements.length) return;

       for (const pElement of fieldElements) {
           if (pElement.classList.contains(className)) {
               pElement.textContent = value;
               pElement.classList.add(classColourSelector);
               break;
           }
       }
   }
   ```

2. A second defensive check was added in `batchCardUtils.js` to fix another error within the `getCardFieldElements` function.
   The card elements are now queried **only if the element exists**, preventing errors caused by attempting to query a non-existent element.

   ```js
   // batchCardUtils.js

   export function getCardFieldElements(cardElement) {
       checkIfHTMLElement(cardElement, "Card Element", true);

       if (cardElement) { // Defensive check added
         .... rest of code here
       }
   }
   ```

**Workflow**

1. On the user’s first setup, generate a 2FA recovery code.
2. The verification form will now appear automatically (without requiring a page refresh).
3. The form remains visible until the user successfully verifies their 2FA code.